Updating Section component
