+++ src/components/Section/index.js
@@ -1,0 +1,3 @@
+import Section from './Section';
+
+export default Section;
